Moved the fetching of products to FilterAndDisplay
